requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.26.0
python-dotenv==0.19.0
spotipy==2.19.0
beautifulsoup4==4.10.0
flask-socketio==5.1.1
gunicorn==20.1.0 